Introduce new JS AST metadata property
Add a new metadata property to JsName that points from
local import alias in inline function to local declaration.
This is required when inlining public function from current module
that calls another function from the same module. In this case
we want to avoid "importing" declaration we already having locally.